1.新增ErrorLog的 .po & .dao 2.修改Order 與 Product 中價錢的資料型態(BigDecimal)

+++ product-repos/src/main/java/team/takoyaki/repos/po/OrderEntity.java
@@ -1,6 +1,7 @@
 package team.takoyaki.repos.po;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 import javax.persistence.Column;
@@ -34,7 +35,7 @@ public class OrderEntity implements Serializable{
 	
 	@Column(name = "TOTAL")
 	@NotNull
-	private Integer total;
+	private BigDecimal total;
 	
 	@Column(name = "AMOUNT")
 	@NotNull
